# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Icategory": {
			"properties": {
				"categoryName": {
					"type": "string"
				}
			},
			"required": [
				"categoryName"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Ibrand": {
			"properties": {
				"brandName": {
					"type": "string"
				}
			},
			"required": [
				"brandName"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Iproduct": {
			"properties": {
				"name": {
					"properties": {
						"ar": {
							"type": "string"
						},
						"en": {
							"type": "string"
						}
					},
					"required": [
						"ar",
						"en"
					],
					"type": "object"
				},
				"price": {
					"type": "number",
					"format": "double"
				},
				"icon": {
					"type": "string"
				},
				"isPremium": {
					"type": "boolean"
				},
				"stock": {
					"type": "number",
					"format": "double"
				},
				"categories": {
					"items": {
						"$ref": "#/definitions/Icategory"
					},
					"type": "array"
				},
				"brandName": {
					"$ref": "#/definitions/Ibrand"
				}
			},
			"required": [
				"name",
				"price",
				"icon",
				"isPremium",
				"stock",
				"categories",
				"brandName"
			],
			"type": "object",
			"additionalProperties": true
		},
...
	},
	"info": {
		"title": "mini-project-2-ecomeernenrenren-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/product": {
			"get": {
				"operationId": "GetProducts",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Iproduct"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddAProduct",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Iproduct"
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"price": {
									"type": "number",
									"format": "double"
								},
								"arabicName": {
									"type": "string"
								},
								"englishName": {
									"type": "string"
								},
								"icon": {
									"type": "string"
								},
								"brandName": {
									"type": "string"
								},
								"categories": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"isPremium": {
									"type": "boolean"
								},
								"stock": {
									"type": "number",
									"format": "double"
								}
							},
							"title": "AddAProductBody",
							"type": "object",
							"required": [
								"price",
								"arabicName",
								"englishName",
								"icon",
								"brandName",
								"categories",
								"isPremium",
								"stock"
							]
						}
					}
				]
			}
		},
		"/product/{_id}": {
			"put": {
				"operationId": "UpdateProduct",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Iproduct"
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"price": {
									"type": "number",
									"format": "double"
								},
								"arabicName": {
									"type": "string"
								},
								"englishName": {
									"type": "string"
								},
								"icon": {
									"type": "string"
								},
								"categories": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"isPremium": {
									"type": "boolean"
								},
								"stock": {
									"type": "number",
									"format": "double"
								}
							},
							"title": "UpdateProductBody",
							"type": "object",
							"required": [
								"price",
								"arabicName",
								"englishName",
								"icon",
								"categories",
								"isPremium",
								"stock"
							]
						}
					}
				]
			}
		},
		"/product/deleteCategory/{_id}": {
			"delete": {
				"operationId": "DeleteItem",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Iproduct"
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"categoryName": {
									"type": "string"
								}
							},
							"title": "DeleteItemBody",
							"type": "object",
							"required": [
								"categoryName"
							]
						}
					}
				]
			}
		},
		"/product/restock/{_id}": {
			"put": {
				"operationId": "ReStockProduct",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"inputStock": {
									"type": "number",
									"format": "double"
								}
							},
							"title": "ReStockProductBody",
							"type": "object",
							"required": [
								"inputStock"
							]
						}
					}
				]
			}
		},
		"/product/deleteProduct/{_id}": {
			"delete": {
				"operationId": "DeleteProduct",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "isLoggedIn",
			"name": "isLoggedIn",
			"in": "query"
		},
		"tsoa_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://swagger.io/api/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	}
}